pyproject.toml
[project]
name = "mc-bridge"
version = "0.1.0"
description = "Monte Carlo Bridge - Local connector for MC SaaS to Snowflake"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "rumps>=0.4.0",
    "snowflake-connector-python[secure-local-storage]>=3.6.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "httpx>=0.26.0",
    "py2app>=0.28.0",
    "ruff>=0.1.0",
]

[project.scripts]
mc-bridge = "mc_bridge.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mc_bridge"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pillow>=12.1.1",
    "pyinstaller>=6.19.0",
    "pytest>=9.0.2",
]

